fix: memory-leak-like issue  in client when adding attesters
