+++ packages/nilai-common/src/nilai_common/logger.py
@@ -33,7 +33,3 @@ def setup_logger(
         logger.addHandler(file_handler)
 
     return logger
-
-
-# Default logger instance
-default_logger = setup_logger("nilai")
